Add cabal.project to allow building with cabal
The posix-waitpid and linux-ptrace dependendencies on Hackage are quite
outdated which is why `stack.yaml` has them as git dependencies instead.

When trying to build with cabal we just get a bunch of solver errors. So I
fix this by adding the cabal equivalent of stack.yaml, cabal.project, with
the git dependencies.